# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Suv yetkazib berish API",
        "description": "Mobil ilovalar uchun suv yetkazib berish tizimining rasmiy API hujjatlari. Barcha endpointlar Uzbek tilida izohlangan bo‘lib, mijoz (user) va xodim (director/courier) ilovalari uchun mo‘ljallangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Asosiy API server (Laravel backend ichidagi /api prefiksi bilan)"
        }
    ],
    "paths": {
...
        "/v1/employees/orders": {
            "get": {
                "tags": [
                    "Xodim – Buyurtmalar"
                ],
                "summary": "Kompaniya buyurtmalarini ko‘rish (direktor/kuryer)",
                "description": "Direktor barcha kompaniya buyurtmalarini status bo‘yicha filtrlashi mumkin. Kuryer faqat o‘zi olishi mumkin bo‘lgan va o‘zi biriktirilgan aktiv buyurtmalarni ko‘radi.",
                "operationId": "e9bbdf195ca4b3e96d310cccaec3ca4d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Faqat direktor uchun: buyurtma statusi bo‘yicha filter (pending, qabul_qilindi, yetkazilmoqda, yetkazildi, canceled)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "pending"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtmalar ro‘yxati muvaffaqiyatli qaytarildi"
                    },
                    "403": {
                        "description": "Ruxsatsiz foydalanuvchi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/orders/{id}/approve": {
            "post": {
                "tags": [
                    "Xodim – Buyurtmalar"
                ],
                "summary": "Direktor tomonidan buyurtmani tasdiqlash",
                "description": "Holati pending bo‘lgan buyurtmani direktor qabul qiladi (`qabul_qilindi`). Shundan so‘ng kuryer uni olish imkoniyatiga ega bo‘ladi.",
                "operationId": "251b6aef60a1a5ca2401df4ee0f09a30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma tasdiqlandi"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki tasdiqlab bo‘lmaydi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/orders/{id}/take": {
            "post": {
                "tags": [
                    "Xodim – Buyurtmalar"
                ],
                "summary": "Kuryer tomonidan buyurtmani qabul qilish",
                "description": "Holati `qabul_qilindi` bo‘lgan va hali kuryer biriktirilmagan buyurtmani kuryer (yoki direktor) o‘ziga oladi va holat `yetkazilmoqda` ga o‘zgaradi.",
                "operationId": "a9bf33b98c5c2be89f71e268715e549e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma kuryerga biriktirildi"
                    },
                    "403": {
                        "description": "Kuryer uchun ruxsat yo‘q"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki allaqachon band qilingan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/orders/{id}/complete": {
            "post": {
                "tags": [
                    "Xodim – Buyurtmalar"
                ],
                "summary": "Buyurtmani yakunlash (yetkazildi deb belgilash)",
                "description": "Kuryer o‘zi biriktirilgan va holati `yetkazilmoqda` bo‘lgan buyurtmani yetkazildi deb belgilaydi.",
                "operationId": "074dfb91134a6bd95d5b3c1b36d438ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma muvaffaqiyatli yakunlandi"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki kuryerga tegishli emas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Xodim – Buyurtmalar"
                ],
                "summary": "Direktor tomonidan buyurtmani bekor qilish",
                "description": "Direktor pending, qabul_qilindi yoki yetkazilmoqda holatidagi buyurtmani bekor qilishi mumkin. Sabab `reason` maydonida yuboriladi.",
                "operationId": "16103eb348a26d0437fda16bc46d5e9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "description": "Bekor qilish sababi (maks. 255 belgi)",
                                        "type": "string",
                                        "example": "Mijoz buyurtmani bekor qildi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Buyurtma direktor tomonidan bekor qilindi"
                    },
                    "403": {
                        "description": "Faqat direktor ushbu amalni bajara oladi"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki bekor qilib bo‘lmaydi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer token bilan autentifikatsiya. Format: \\\"Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}